pyproject.toml
[project]
name = "quaestor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sec-edgar-downloader>=5.1.0",
    "langgraph>=1.1.3",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "qdrant-client>=1.17.1",
    "langchain-qdrant>=1.1.0",
    "fastembed>=0.8.0",
    "presidio-analyzer>=2.2.362",
    "presidio-anonymizer>=2.2.362",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "reportlab>=4.4.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quaestor"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
